requirements.txt
aenum==2.0.8
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
enum==0.4.6
freezegun==0.3.9
idna==2.6
py==1.4.34
pytest==3.2.2
pytest-localserver==0.3.7
python-dateutil==2.6.1
requests==2.18.4
six==1.10.0
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12.2
